requirements.txt
discord.py==2.3.1
python-dotenv==1.0.0
aiohttp==3.8.4
requests==2.31.0
